#30 Finished configuring core and login modules
