+++ frontend/src/app/app-routing.module.ts
@@ -1,48 +1,18 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './core/home/home.component';
-import { LoginComponent } from './core/login/login.component';
 import { UserGuard } from './core/auth/user.guard';
-import { RoleGuard } from "./core/guards/role-guard";
 import { LoginGuard } from "./core/auth/login.guard";
 
 const routes: Routes = [
   {
-    path: '',
-    component: HomeComponent,
-    canActivate: [UserGuard],
-  },
-  {
     path: 'login',
-    component: LoginComponent,
+    loadChildren: () => import('src/app/modules/login/login.module').then(m => m.LoginModule),
     canActivate: [LoginGuard]
   },
   {
-    path: 'invoices',
-    loadChildren: () => import('src/app/modules/core/invoices/invoices.module').then(m => m.InvoicesModule),
+    path: '',
+    loadChildren: () => import('src/app/modules/core/core.module').then(m => m.CoreModule),
     canActivate: [UserGuard],
-  },
-  {
-    path: 'suppliers',
-    loadChildren: () => import('src/app/modules/core/suppliers/suppliers.module').then(m => m.SuppliersModule),
-    canActivate: [UserGuard, RoleGuard],
-    data: {
-      allowedRole: 'ADMIN'
-    }
-  },
-  {
-    path: 'products',
-    loadChildren: () => import('src/app/modules/core/products/products.module').then(m => m.ProductsModule),
-    canActivate: [UserGuard],
-  },
-  {
-    path: 'delivery',
-    loadChildren: () => import('src/app/modules/core/delivery/delivery.module').then(m => m.DeliveryModule),
-    canActivate: [UserGuard],
-  },
-  {
-    path: '**',
-    component: HomeComponent,
   }
 ];
