+++ frontend/src/app/app.component.html
@@ -1,12 +1,1 @@
-@if ((userService.isLoggingIn$ | async) === false) {
-<div class="flex overflow-y-hidden h-screen">
-  <app-nav></app-nav>
-  <div class="w-full overflow-y-scroll">
-    <router-outlet></router-outlet>
-  </div>
-</div>
-} @else {
-  <div class="w-full h-full overflow-y-scroll">
-    <router-outlet></router-outlet>
-  </div>
-}
+<router-outlet></router-outlet>
