+++ frontend/src/app/modules/core/core-routing.module.ts
@@ -1,0 +1,53 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import {UserGuard} from "../../core/auth/user.guard";
+import {RoleGuard} from "../../core/guards/role-guard";
+import {HomeComponent} from "../../core/home/home.component";
+import {CoreComponent} from "./core.component";
+
+const routes: Routes = [
+  {
+    path: '',
+    component: CoreComponent,
+    children: [
+      {
+        path: '',
+        component: HomeComponent,
+        canActivate: [UserGuard]
+      },
+      {
+        path: 'invoices',
+        loadChildren: () => import('src/app/modules/core/invoices/invoices.module').then(m => m.InvoicesModule),
+        canActivate: [UserGuard],
+      },
+      {
+        path: 'suppliers',
+        loadChildren: () => import('src/app/modules/core/suppliers/suppliers.module').then(m => m.SuppliersModule),
+        canActivate: [UserGuard, RoleGuard],
+        data: {
+          allowedRole: 'ADMIN'
+        }
+      },
+      {
+        path: 'products',
+        loadChildren: () => import('src/app/modules/core/products/products.module').then(m => m.ProductsModule),
+        canActivate: [UserGuard],
+      },
+      {
+        path: 'delivery',
+        loadChildren: () => import('src/app/modules/core/delivery/delivery.module').then(m => m.DeliveryModule),
+        canActivate: [UserGuard],
+      },
+      {
+        path: '**',
+        component: HomeComponent,
+      }
+    ]
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class CoreRoutingModule { }
