+++ frontend/src/app/modules/core/core.component.html
@@ -1,0 +1,6 @@
+<div class="flex overflow-y-hidden h-screen">
+  <app-nav></app-nav>
+  <div class="w-full overflow-y-scroll">
+    <router-outlet></router-outlet>
+  </div>
+</div>
