+++ frontend/src/app/modules/core/core.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-core',
+  templateUrl: './core.component.html',
+  styleUrl: './core.component.scss'
+})
+export class CoreComponent {
+
+}
