--- frontend/src/app/modules/core/core/core.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { CoreRoutingModule } from './core-routing.module';
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    CoreRoutingModule
-  ]
-})
-export class CoreModule { }
