+++ frontend/src/app/modules/login/login.component.html
@@ -1,0 +1,3 @@
+<div class="w-full h-full overflow-y-scroll">
+  <router-outlet></router-outlet>
+</div>
